move 'slice' tests to separate file
